refactor: 응집도가 낮은 함수 개선
